# List performance improvent by Union-typed tail