@oxinabox 2020-09-21
Currently we are stongly considering deprecating DataStructure.jl's LinkedLists in favour of FunctionalCollection's equivelent (`plist`).

You might want to make this PR over there:
https://github.com/JuliaCollections/FunctionalCollections.jl/blob/2dd9b36fd6e637c866011363ba6b7097345d0df4/src/PersistentList.jl#L8